pyproject.toml
[project]
name = "sk30-1st-4team"
version = "0.1.0"
description = "전기차 충전 인프라 대시보드"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jpype1>=1.6.0",
    "jupyter>=1.1.1",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
    "wordcloud>=1.9.6",
    "pydantic>=2.0.0",
    "pandera[pandas]>=0.13.0",
    "xmltodict>=0.13.0",
    "folium>=0.20.0",
    "beautifulsoup4>=4.14.3",
    "streamlit-folium>=0.27.1",
    "plotly>=6.7.0",
    "requests>=2.33.1",
]
